requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==2.3.0
Django==5.1.7
django-mathfilters==1.0.0
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-barcode==0.15.1
reportlab==4.4.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.1
whitenoise==6.9.0
